requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
PyPDF2
langchain-core
dotenv
faiss-cpu
pypdf
python-docx 
pytesseract 
pdf2image 
pillow